fix: Removed async_iterator file
