+++ llama-index-integrations/llms/llama-index-llms-ai21/tests/BUILD
@@ -1,10 +1,3 @@
-python_tests(
-  name="tests",
-  sources=["async_iterator.py", "test_llms_ai21.py"],
-  dependencies=[
-  ],
-)
-
 python_sources(
     name="tests0",
 )
